feat: refactor Spotify integration by moving access token logic to a separate action and updating now playing widget
